+++ substrata-kotlin/src/androidTest/java/com/segment/analytics/substrata/kotlin/EngineTests.kt
@@ -22,7 +22,6 @@ class EngineTests {
     @After
     fun tearDown() {
         scope.release()
-        JSShared.reset()
     }
 
     @Test
